backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
redis
minio
python-multipart
python-dotenv
